[fix] #112 엔티티 통합 및 DTO 분리
