+++ SniffMeet/SniffMeet/Source/Entity/DogTraits/Keyword.swift
@@ -1,0 +1,16 @@
+//
+//  Keyword.swift
+//  SniffMeet
+//
+//  Created by Kelly Chui on 11/28/24.
+//
+
+import Foundation
+
+enum Keyword: String, Codable  {
+    case energetic = "활발한"
+    case smart = "똑똑한"
+    case friendly = "친화력 좋은"
+    case shy = "소심한"
+    case independent = "독립적인"
+}
