+++ SniffMeet/SniffMeet/Source/Entity/DogTraits/Sex.swift
@@ -1,0 +1,13 @@
+//
+//  Sex.swift
+//  SniffMeet
+//
+//  Created by Kelly Chui on 11/28/24.
+//
+
+import Foundation
+
+enum Sex: String, Codable {
+    case male = "남"
+    case female = "여"
+}
