+++ SniffMeet/SniffMeet/Source/Entity/DogTraits/Size.swift
@@ -1,0 +1,14 @@
+//
+//  Size.swift
+//  SniffMeet
+//
+//  Created by Kelly Chui on 11/28/24.
+//
+
+import Foundation
+
+enum Size: String, Codable {
+    case small = "소형"
+    case medium = "중형"
+    case big = "대형"
+}
