+++ SniffMeet/SniffMeet/Source/Supabase/Auth/Entity/SupabaseUser.swift
@@ -9,12 +9,8 @@ import Foundation
 
 struct SupabaseUser: Encodable {
     var userID: UUID
-    // var name: String
-    // var profileImageURL: URL
 
     init(from response: SupabaseUserResponse) {
         self.userID = response.id
-        // self.name = response.userMetadata.name
-        // self.profileImageURL = response.userMetadata.profileImageURL
     }
 }
